requirements.txt
torch -i https://download.pytorch.org/whl/cu118
torchvision -i https://download.pytorch.org/whl/cu118
torchaudio -i https://download.pytorch.org/whl/cu118
platformdirs
pillow
opencv-python
PyYAML
transformers
gdown
scipy
accelerate>=0.26.0
